refactor: move types, rename hook
